// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/marketing-attribution-day-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface MarketingAttributionDayHttpResponse {
    date?: string;
    premiumStarts?: number;
    signups?: number;
    visits?: number;
}