// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/marketing-attribution-range-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MarketingAttributionDayHttpResponse } from './marketing-attribution-day-http-response';
import { MarketingAttributionSummaryHttpResponse } from './marketing-attribution-summary-http-response';

export interface MarketingAttributionRangeHttpResponse {
    byDay?: Array<MarketingAttributionDayHttpResponse>;
    current?: MarketingAttributionSummaryHttpResponse;
    eventTotal?: number;
    fromUtc?: string;
    previous?: MarketingAttributionSummaryHttpResponse;
    previousFromUtc?: string;
    toUtc?: string;
}