// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/marketing-attribution-recent-event-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface MarketingAttributionRecentEventHttpResponse {
    anonymousId?: string;
    buildVersion?: string | null;
    eventType?: string;
    landingPath?: string;
    occurredAtUtc?: string;
    referrerHost?: string | null;
    sessionId?: string;
    utmCampaign?: string | null;
    utmContent?: string | null;
    utmMedium?: string | null;
    utmSource?: string | null;
    utmTerm?: string | null;
}