// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/marketing-attribution-summary-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MarketingAttributionBreakdownHttpResponse } from './marketing-attribution-breakdown-http-response';
import { MarketingAttributionRecentEventHttpResponse } from './marketing-attribution-recent-event-http-response';

export interface MarketingAttributionSummaryHttpResponse {
    anonymousVisitors?: number;
    attributedEvents?: number;
    attributedVisits?: number;
    events?: number;
    generatedAtUtc?: string;
    lastEventAtUtc?: string | null;
    organicEvents?: number;
    organicVisits?: number;
    premiumRatePercent?: number;
    premiumStarts?: number;
    recentEvents?: Array<MarketingAttributionRecentEventHttpResponse>;
    sessions?: number;
    signupRatePercent?: number;
    signups?: number;
    topCampaigns?: Array<MarketingAttributionBreakdownHttpResponse>;
    topSources?: Array<MarketingAttributionBreakdownHttpResponse>;
    visits?: number;
    windowHours?: number;
}